respaldo

+++ libs/create-apps-test.sh
@@ -14,6 +14,11 @@ PATH_CONFIG_NGINX_APPS=$3
 
 sudo mkdir -p $PATH_CONFIG_NGINX_APPS
 
+if ! hash pm2 2>/dev/null; then
+    echo "requires pm2"
+    exit 0
+fi
+
 pm2 delete all
 sudo rm -rf $PATH_APPS
 
@@ -22,6 +27,8 @@ for (( i=$START; i<=$END; i++ ))
 do
     APP_NAME="app-00${i}"
     APP_POST="300${i}"
-    source $PWD/libs/create-app-test.sh $PATH_APPS $APP_NAME $APP_POST
-    source $PWD/libs/config-nginx-app.sh $PATH_CONFIG_NGINX_APPS $APP_NAME $APP_POST
+    APP_PATH="${PATH_APPS}/${APP_NAME}"
+    PATH_CONFIG_FILE="${PATH_CONFIG_NGINX_APPS}/${APP_NAME}.conf"
+    source $PWD/libs/create-app-test.sh $APP_PATH $APP_NAME $APP_POST
+    source $PWD/libs/config-nginx-app.sh $PATH_CONFIG_FILE $APP_NAME $APP_POST
 done
